Revert "Updated media service"
