--- src/core/services/devices/index.ts
@@ -1,1 +1,0 @@
-export class DevicesService {}
