--- src/core/services/stream/video/index.ts
@@ -1,11 +1,0 @@
-import { BaseService } from '../base';
-
-export class VideoService extends BaseService {
-  constructor() {
-    super({ type: 'video' });
-  }
-
-  public override async get() {
-    super.get({ video: true });
-  }
-}
